refactor: move `MultipartParseError` to `falcon.errors`, other misc cleanup
