+++ tests/test_inspect.py
@@ -5,18 +5,16 @@ import sys
 import _inspect_fixture as i_f
 import pytest
 
+import falcon
+import falcon.asgi
 from falcon import inspect, routing
 
 
 def get_app(asgi, cors=True, **kw):
     if asgi:
-        from falcon.asgi import App as AsyncApp
-
-        return AsyncApp(cors_enable=cors, **kw)
+        return falcon.asgi.App(cors_enable=cors, **kw)
     else:
-        from falcon import App
-
-        return App(cors_enable=cors, **kw)
+        return falcon.App(cors_enable=cors, **kw)
 
 
 def make_app():
